1) An experimental parallelization of NEB images has been implemented. This required a deep modification of the parallelism in PWscf: there are two new communicators (intra_image_comm and inter_image_comm) and the existing "pool" communicators (intra_pool_comm and inter_pool_comm) are now vectors of length given by the number of parallel images. #ifdef __PARA is no longer needed because all "parallel" variables are always initialized for a serial run and all parallel routines are, in the case of a serial run, dummy routines. The wrappers to MPI routines used only by PWscf are in the PW/para.f90 file. The others (mp_***) are in the Modules/mp.f90. All explicit referencies to mpif.h should be replaced by an "USE parallel_include" (in a serial run parallel_include is simply a dummy module).
2) The extrapolation of both potential and wavefunctions has been rewritten in
order to be smarter than before: on the basis of the required extrapolation
order, on the basis of the history and on the basis of which files are really
present on the disk, the algorithm chooses the extrapolation order.
All the algorithms in which ions are moved can use the extrapolation.

These are both unstable features: I need the help of everybody to test them.

C.S.